refactors downloader code for macOS & iOS
More consistent code across platforms.
